all requested modifications were made
