+++ react-app/src/app/screens/Game/components/Board/index.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
 
 import Square from '../Square';
@@ -10,7 +10,7 @@ class Board extends Component {
 
   render() {
     return (
-      <React.Fragment>
+      <Fragment>
         <div className={style.boardRow}>
           {this.renderSquare(0)}
           {this.renderSquare(1)}
@@ -26,7 +26,7 @@ class Board extends Component {
           {this.renderSquare(7)}
           {this.renderSquare(8)}
         </div>
-      </React.Fragment>
+      </Fragment>
     );
   }
 }
