+++ react-app/src/app/screens/Game/components/Board/styles.scss
@@ -1,3 +1,5 @@
+@import '../../../../../scss/colors';
+
 .board-row:after {
   clear: both;
   content: "";
