+++ react-app/src/app/screens/Game/components/Square/index.js
@@ -3,10 +3,10 @@ import PropTypes from 'prop-types';
 
 import style from './styles.scss';
 
-function Square(props) {
+function Square({ value, onClick }) {
   return (
-    <button className={style.square} onClick={props.onClick}>
-      {props.value}
+    <button className={style.square} onClick={onClick}>
+      {value}
     </button>
   );
 }
