+++ react-app/src/app/screens/Game/components/Square/styles.scss
@@ -1,5 +1,7 @@
+@import '../../../../../scss/colors';
+
 .square {
-  background: #fff;
+  background: $white;
   border: 1px solid #999;
   float: left;
   font-size: 24px;
@@ -8,7 +10,7 @@
   line-height: 34px;
   margin-right: -1px;
   margin-top: -1px;
-  padding: 0;
+  padding: 0; 
   text-align: center;
   width: 34px;
