+++ react-app/src/app/screens/Game/index.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 
+import calculateWinner from './utils';
 import Board from './components/Board';
 import style from './styles.scss';
 
@@ -14,7 +15,7 @@ class Game extends Component {
     const history = this.state.history.slice(0, this.state.stepNumber + 1);
     const current = history[history.length - 1];
     const squares = current.squares.slice();
-    if (this.calculateWinner(squares) || squares[i]) {
+    if (calculateWinner(squares) || squares[i]) {
       return;
     }
     squares[i] = this.state.xIsNext ? 'X' : 'O';
@@ -24,36 +25,28 @@ class Game extends Component {
       xIsNext: !this.state.xIsNext
     });
   };
+
+  createHistoryButton(move) {
+    const desc = move ? `Go to move # ${move}` : 'Go to game start';
+    return (
+      <li key={move.toString()}>
+        <button onClick={() => this.jumpTo(move)}>{desc}</button>
+      </li>
+    );
+  }
+
   jumpTo = step => {
     this.setState({
       stepNumber: step,
       xIsNext: step % 2 === 0
     });
   };
-  calculateWinner = squares => {
-    const lines = [[0, 1, 2], [3, 4, 5], [6, 7, 8], [0, 3, 6], [1, 4, 7], [2, 5, 8], [0, 4, 8], [2, 4, 6]];
-    for (let i = 0; i < lines.length; i += 1) {
-      const [a, b, c] = lines[i];
-      if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-        return squares[a];
-      }
-    }
-    return null;
-  };
 
   render() {
     const history = this.state.history;
     const current = history[this.state.stepNumber];
-    const winner = this.calculateWinner(current.squares);
-
-    const moves = history.map((step, move) => {
-      const desc = move ? `Go to move # ${move}` : 'Go to game start';
-      return (
-        <li key={move.toString()}>
-          <button onClick={() => this.jumpTo(move)}>{desc}</button>
-        </li>
-      );
-    });
+    const winner = calculateWinner(current.squares);
+    const moves = history.map((step, move) => this.createHistoryButton(move));
 
     let status;
     if (winner) {
@@ -65,7 +58,7 @@ class Game extends Component {
     return (
       <div className={style.game}>
         <div className={style.gameBoard}>
-          <Board squares={current.squares} onClick={i => this.handleClick(i)} />
+          <Board squares={current.squares} onClick={this.handleClick} />
         </div>
         <div className={style.gameInfo}>
           <div>{status}</div>
