+++ react-app/src/app/screens/Game/styles.scss
@@ -1,3 +1,5 @@
+@import '../../../../src/scss/colors.scss';
+
 .game {
   display: flex;
   flex-direction: row;
