+++ react-app/src/app/styles.scss
@@ -1,3 +1,5 @@
+@import '../scss/colors.scss';
+
 .app {
   text-align: center;
 }
@@ -8,8 +10,8 @@
 }
 
 .app-header {
-  background-color: #222;
-  color: white;
+  background-color: $mine-shaft;
+  color: $white;
   height: 150px;
   padding: 20px;
 }
