+++ react-app/src/index.js
@@ -1,10 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import App from 'app';
 
 import './scss/index.scss';
 import registerServiceWorker from './registerServiceWorker';
-
-import App from '~components/App'; // eslint-disable-line import/first
 
 ReactDOM.render(<App />, document.getElementById('root'));
 registerServiceWorker();
