+++ react-app/src/scss/colors.scss
@@ -1,0 +1,2 @@
+$white: #fff;
+$mine-shaft: #222;
