+++ react-app/src/scss/index.scss
@@ -1,3 +1,5 @@
+@import 'colors';
+
 body {
   padding: 0;
   font: 14px "Century Gothic", Futura, sans-serif;
@@ -6,4 +8,3 @@ body {
 ol, ul {
   padding-left: 30px;
 }
-
